+++ waffle_hub/dataset/adapter/__init__.py
@@ -1,6 +1,15 @@
-from .autocare_dlt import export_autocare_dlt
-from .coco import export_coco
-from .transformers import export_transformers
-from .yolo import export_yolo
+from .autocare_dlt import export_autocare_dlt, import_autocare_dlt
+from .coco import export_coco, import_coco
+from .transformers import export_transformers, import_transformers
+from .yolo import export_yolo, import_yolo
 
-__all__ = ["export_yolo", "export_coco", "export_transformers", "export_autocare_dlt"]
+__all__ = [
+    "export_yolo",
+    "export_coco",
+    "export_transformers",
+    "export_autocare_dlt",
+    "import_autocare_dlt",
+    "import_coco",
+    "import_transformers",
+    "import_yolo",
+]
